refactor: Create subfolders for action related classes
