+++ src/main/java/com/onarandombox/MultiverseCore/MultiverseCore.java
@@ -63,7 +63,7 @@ import com.onarandombox.MultiverseCore.teleportation.SimpleBlockSafety;
 import com.onarandombox.MultiverseCore.teleportation.SimpleLocationManipulation;
 import com.onarandombox.MultiverseCore.teleportation.SimpleSafeTTeleporter;
 import com.onarandombox.MultiverseCore.utils.PermissionsTool;
-import com.onarandombox.MultiverseCore.utils.PlayerActionChecker;
+import com.onarandombox.MultiverseCore.utils.actioncheck.PlayerActionChecker;
 import com.onarandombox.MultiverseCore.utils.TestingMode;
 import com.onarandombox.MultiverseCore.utils.UnsafeCallWrapper;
 import com.onarandombox.MultiverseCore.utils.metrics.MetricsConfigurator;
