+++ src/main/java/com/onarandombox/MultiverseCore/api/MVCore.java
@@ -15,7 +15,7 @@ import com.onarandombox.MultiverseCore.teleportation.SimpleBlockSafety;
 import com.onarandombox.MultiverseCore.teleportation.SimpleLocationManipulation;
 import com.onarandombox.MultiverseCore.teleportation.SimpleSafeTTeleporter;
 import com.onarandombox.MultiverseCore.utils.PermissionsTool;
-import com.onarandombox.MultiverseCore.utils.PlayerActionChecker;
+import com.onarandombox.MultiverseCore.utils.actioncheck.PlayerActionChecker;
 import com.onarandombox.MultiverseCore.utils.UnsafeCallWrapper;
 
 /**
