+++ src/main/java/com/onarandombox/MultiverseCore/listeners/MVPlayerListener.java
@@ -17,10 +17,10 @@ import com.onarandombox.MultiverseCore.MultiverseCore;
 import com.onarandombox.MultiverseCore.api.MVWorld;
 import com.onarandombox.MultiverseCore.api.MVWorldManager;
 import com.onarandombox.MultiverseCore.event.MVRespawnEvent;
-import com.onarandombox.MultiverseCore.utils.action.ActionResponse;
-import com.onarandombox.MultiverseCore.utils.ActionCheckResult;
-import com.onarandombox.MultiverseCore.utils.PlayerActionChecker;
-import com.onarandombox.MultiverseCore.utils.action.ActionResult;
+import com.onarandombox.MultiverseCore.api.action.ActionResponse;
+import com.onarandombox.MultiverseCore.utils.actioncheck.ActionCheckResult;
+import com.onarandombox.MultiverseCore.utils.actioncheck.PlayerActionChecker;
+import com.onarandombox.MultiverseCore.api.action.ActionResult;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.Material;
